// src/lib/orval/generated/model/experiencesControllerListStudentsWithAccess200DataItem.ts
/**
 * Generated by orval v7.13.1 🍺
 * Do not edit manually.
 * Gradegy API
 * Interactive documentation for the Gradegy API
 * OpenAPI spec version: 1.0
 */
import type { ExperiencesControllerListStudentsWithAccess200DataItemUser } from './experiencesControllerListStudentsWithAccess200DataItemUser';
import type { UserAssignment } from './userAssignment';

export type ExperiencesControllerListStudentsWithAccess200DataItem = {
  user?: ExperiencesControllerListStudentsWithAccess200DataItemUser;
  assignments?: UserAssignment[];
  /** Whether the student has completed the experience. */
  completed?: boolean;
  /**
   * Completion timestamp when the experience was finished by the student.
   * @nullable
   */
  completedAt?: string | null;
};